README: finalising GfM migration
And this concludes the switch to GitHub-flavoured Markdown! This will sync the content between the main GitHub repo and GitHub Pages so you won't have to see any annoying style tags (although GfM is quite limited, that should be enough for our purpose, documenting a device). Also with this, scrollable code on smaller screens